Add files via upload
Project created to put into practice the content learned in the Rafa OutSystems Academy course. This project allows the user to create lists of Subjects, Topics and Cards for study, and view the Cards created by marking whether they got the correct answer or not.
In this project, several elements of the outsystems platform, database creation and manipulation, good development practices and much more were used.